Add CI (GitHub Actions) and ruff lint, fix everything it found
Added .github/workflows/ci.yml: pytest across Python 3.9/3.11/3.12 on
ubuntu-latest (the only platform the loopback-binding model needs zero
extra setup on), plus a separate ruff lint job. Runs on push to main,
on pull requests, and on demand.

Added an explicit ruff config (pyproject.toml) rather than relying on
whatever ruff's default selection happens to be in a given version:
pyflakes/pycodestyle correctness rules, import sorting, pyupgrade (kept
3.9-safe via target-version), bugbear, comprehension and simplification
suggestions, blind-except detection, and ruff's own correctness rules.

Fixed everything it flagged: unsorted imports and __all__ lists,
unnecessary quoted type annotations (safe project-wide since every
file already has `from __future__ import annotations`), dict() calls
rewritten as literals, a redundant int() around round() (which already
returns int), unused unpacked variables in tests, two pytest.raises(Exception)
assertions narrowed to the actual exception type raised (InvalidTag),
and two real asyncio dangling-task bugs: forward_downstream_to_upstream's
task and the dashboard's background run task were never referenced
after creation, meaning asyncio could garbage-collect either mid-run.
Both now keep an explicit reference for their whole lifetime.

The two remaining `except Exception` catches are intentional API/task
boundaries (untrusted YAML parsing, a background task's only path to
reporting failure) and are suppressed inline with a comment explaining
why, rather than narrowed to specific types that would defeat the point
of catching everything there.

+++ anontestlab/metrics/stats.py
@@ -38,7 +38,7 @@ def tpr_at_fpr(
         if n_impostor == 0:
             result[fpr] = float("nan")
             continue
-        k = max(1, int(round(fpr * n_impostor)))
+        k = max(1, round(fpr * n_impostor))
         k = min(k, n_impostor)
         threshold = impostor[k - 1]
         tpr = float(np.mean(true >= threshold)) if len(true) else float("nan")
